requirements.txt
annotated-types==0.7.0
anyio==4.8.0
atlassian-python-api==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
cssselect==1.2.0
deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-mcp==0.3.4
fastmcp==2.3.4
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
identify==2.6.7
idna==3.10
iniconfig==2.0.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jinja2==3.1.6
jmespath==1.0.1
keyring==25.6.0
lxml==5.3.1
markdown==3.7
markdown-it-py==3.0.0
markdown-to-confluence==0.3.0
markdownify==0.14.1
markupsafe==3.0.2
mcp==1.8.1
mdurl==0.1.2
more-itertools==10.7.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
openapi-pydantic==0.5.1
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==4.1.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.1
pymdown-extensions==10.14.3
pysocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
rapidfuzz==3.13.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.14.6
ruff==0.9.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.45.3
thefuzz==0.22.1
tomli==2.2.1
trio==0.29.0
typer==0.15.2
types-cachetools==5.5.0.20240820
types-html5lib==1.1.11.20241018
types-lxml==2025.3.4
types-markdown==3.7.0.20241204
types-python-dateutil==2.9.0.20241206
types-pyyaml==6.0.12.20241230
types-requests==2.32.0.20250301
typing-extensions==4.12.2
urllib3==2.3.0
uv==0.6.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
